requirements.txt
pandas
matplotlib
seaborn
python-dotenv
langchain-community
langchain-chroma
transformers
notebook
psycopg2-binary  # Keep this, ONLY this for psycopg
langchain_ollama
torch
gradio
langchain-postgres  # Make sure this is present